- Made SessionData an interface, and introduced DefaultSessionData with the implementation previous in SessionData
