+++ pippo-session-parent/pippo-session-cookie/src/main/java/ro/pippo/session/cookie/CookieSessionDataStorage.java
@@ -22,6 +22,7 @@ import ro.pippo.core.Response;
 import ro.pippo.core.util.CookieUtils;
 import ro.pippo.session.SerializationSessionDataTranscoder;
 import ro.pippo.session.SessionData;
+import ro.pippo.session.DefaultSessionData;
 import ro.pippo.session.SessionDataStorage;
 import ro.pippo.session.SessionDataTranscoder;
 
@@ -48,7 +49,7 @@ public class CookieSessionDataStorage implements SessionDataStorage {
 
     @Override
     public SessionData create() {
-        return new SessionData();
+        return new DefaultSessionData();
     }
 
     @Override
