+++ pippo-session-parent/pippo-session-ehcache2/src/main/java/ro/pippo/session/ehcache/EhcacheSessionDataStorage.java
@@ -18,6 +18,7 @@ package ro.pippo.session.ehcache;
 import net.sf.ehcache.Cache;
 import net.sf.ehcache.CacheManager;
 import net.sf.ehcache.Element;
+import ro.pippo.session.DefaultSessionData;
 import ro.pippo.session.SessionData;
 import ro.pippo.session.SessionDataStorage;
 
@@ -60,7 +61,7 @@ public class EhcacheSessionDataStorage implements SessionDataStorage {
 
     @Override
     public SessionData create() {
-        return new SessionData();
+        return new DefaultSessionData();
     }
 
     @Override
