+++ pippo-session-parent/pippo-session-infinispan/src/main/java/ro/pippo/session/infinispan/InfinispanSessionDataStorage.java
@@ -17,6 +17,7 @@ package ro.pippo.session.infinispan;
 
 import org.infinispan.Cache;
 import org.infinispan.manager.EmbeddedCacheManager;
+import ro.pippo.session.DefaultSessionData;
 import ro.pippo.session.SessionData;
 import ro.pippo.session.SessionDataStorage;
 
@@ -60,7 +61,7 @@ public class InfinispanSessionDataStorage implements SessionDataStorage {
 
     @Override
     public SessionData create() {
-        return new SessionData();
+        return new DefaultSessionData();
     }
 
     @Override
