+++ pippo-session-parent/pippo-session-jdbc/src/main/java/ro/pippo/session/jdbc/JDBCSessionDataStorage.java
@@ -24,10 +24,7 @@ import javax.sql.DataSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import ro.pippo.core.PippoRuntimeException;
-import ro.pippo.session.SerializationSessionDataTranscoder;
-import ro.pippo.session.SessionData;
-import ro.pippo.session.SessionDataStorage;
-import ro.pippo.session.SessionDataTranscoder;
+import ro.pippo.session.*;
 
 /**
  * SessionDataStorage implementation with JDBC.
@@ -70,7 +67,7 @@ public class JDBCSessionDataStorage implements SessionDataStorage {
 
     @Override
     public SessionData create() {
-        return new SessionData();
+        return new DefaultSessionData();
     }
 
     @Override
