+++ pippo-session-parent/pippo-session-jedis/src/main/java/ro/pippo/session/jedis/JedisSessionDataStorage.java
@@ -17,10 +17,7 @@ package ro.pippo.session.jedis;
 
 import redis.clients.jedis.Jedis;
 import redis.clients.jedis.JedisPool;
-import ro.pippo.session.SerializationSessionDataTranscoder;
-import ro.pippo.session.SessionData;
-import ro.pippo.session.SessionDataStorage;
-import ro.pippo.session.SessionDataTranscoder;
+import ro.pippo.session.*;
 
 /**
  * SessionDataStorage implementation with Jedis.
@@ -69,7 +66,7 @@ public class JedisSessionDataStorage implements SessionDataStorage {
 
     @Override
     public SessionData create() {
-        return new SessionData();
+        return new DefaultSessionData();
     }
 
     @Override
