+++ pippo-session-parent/pippo-session/src/main/java/ro/pippo/session/MemorySessionDataStorage.java
@@ -31,7 +31,7 @@ public class MemorySessionDataStorage implements SessionDataStorage {
 
     @Override
     public SessionData create() {
-        return new SessionData();
+        return new DefaultSessionData();
     }
 
     @Override
